refactor: separado os testes em arquivos individuais
